% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testData.R
\docType{data}
\name{testData}
\alias{testData}
\title{Test data for munging}
\format{An object of class \code{matrix} with 6 rows and 6 columns.}
\usage{
data(testData)
}
\description{
Test data for your munging pleasure.
}
\examples{
data(testData)
mung(testData)
}
\keyword{datasets}
