% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testData.R
\docType{data}
\name{testData2}
\alias{testData2}
\title{Test data for munging}
\format{An object of class \code{data.frame} with 6 rows and 33 columns.}
\usage{
data(testData2)
}
\description{
Test data for your munging pleasure.
}
\examples{
data(testData)
mung(testData2)
}
\keyword{datasets}
